Make clear that lua binary is not enough
It may not be clear that this module requires not a lua compiler as binary but the lua library and header files. For instance Ubuntu has lua50 package and a liblua50-dev package.

By the way, packaging the Lua source code (only 1.2MB) with this package, or giving instructions where to put the files from http://www.lua.org/source/, would simplify installation a lot!