Respect NV OK flag before PV OK flag
See GH #32

The issue is that if you have a scalar with a number in it, and then you
print it for debugging, the scalar will have the side effect of getting
a PV buffer and flipping the PV OK flag on.  Before this change,
push_val will *always* pass that scalar to Lua as a string.

This change just checks the NV flag first, avoiding the above situation